Refactored client to move the initial listener into its own folder. Note - we may have more listeners in the future.
