+++ source/RunClient.py
@@ -1,51 +1,8 @@
 import sys
 from time import sleep
 
-from PodSixNet.Connection import connection, ConnectionListener
-
-class GameClient(ConnectionListener):
-    """ This client connects to a GameServer which will host a cardgame
-
-    Currently the client is incomplete
-
-    The client displays the game state it recieves from the server
-    It validates and submits player actions to the server during the player's turn
-    It submits its score on round or game end
-    """
-
-    def __init__(self):
-        self.name = input("Select a display name: ")
-        interactive = False #wait for the server say its this player's turn
-        connection.Send({"action": "displayName", "name": self.name})
-
-    #######################################
-    ### Network event/message callbacks ###
-    #######################################
-    # built in stuff
-    
-    def Network_connected(self, data):
-        print("Connected to the server")
-    
-    def Network_error(self, data):
-        print('error:', data['error'])
-        connection.Close()
-    
-    def Network_disconnected(self, data):
-        print('Server disconnected')
-        exit()
-
-    # server specific stuff
-    def Network_connectionDenied(self, data):
-        """Server denied the connection, likely due to a game already in progress"""
-        print('Server denied connection request')
-        connection.Close()
-
-    def Network_turnOrder(self, data):
-        """Turn order changed"""
-        print('Turn order is', data['players'])
-
-    ### Player Actions ###
-    # here is where to put any actions triggered by the player that should be sent to the server
+from PodSixNet.Connection import connection
+from client.BaseClientListener import GameClient
 
 
 if len(sys.argv) != 2:
